third_party/WebKit/LayoutTests/fast/dom/getElementsByClassName/non-styled-elements.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../resources/js-test.js"></script>
</head>
<body>
<script type="text/javascript">
description("This test checks to see if getElementsByClassName finds non styled elements.");

var d = document.createElement("div");
d.setAttribute("class", "target");
document.documentElement.appendChild(d);

var e = document.createElementNS("http://foo.com", "div");
e.setAttribute("class", "target");
document.documentElement.appendChild(e);

shouldBe("document.getElementsByClassName('target').length", "2");
</script>
</body>
</html>
