third_party/WebKit/LayoutTests/fast/dom/getElementsByClassName/non-styled-elements-expected.txt

This test checks to see if getElementsByClassName finds non styled elements.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS document.getElementsByClassName('target').length is 2
PASS successfullyParsed is true

TEST COMPLETE
